move logic to server component
