+++ src/components/Oval/OevLost.tsx
@@ -1,6 +1,24 @@
 import { Divider } from "./Divider";
 import { Countup } from "./Countup";
-import { getOevLost } from "@/queries/getOevLost";
+import { OevDataResponse } from "@/app/api/oev-data/route";
+import { getApiRouteUrl } from "@/utils";
+import { headers } from "next/headers";
+
+// should this be an env variable?
+const ONE_DAY_SECONDS = 86_400;
+
+export async function getOevLost() {
+  const host = headers().get("host");
+  const URI = getApiRouteUrl("/api/oev-data", host);
+
+  const response = await fetch(URI, {
+    next: {
+      revalidate: ONE_DAY_SECONDS,
+    },
+  });
+  const res = (await response.json()) as OevDataResponse;
+  return res.oevLost;
+}
 
 export const OevLost = async () => {
   const oevLost = await getOevLost();
